--- src/config/entry.js
@@ -1,8 +1,0 @@
-/**
- * Webpack entry points
- */
-
-export default {
-	'admin-common': ['./scripts/admin.js', './styles/admin.scss'],
-	common: ['./scripts/main.js', './styles/main.scss'],
-};
